pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "boatrace-ai"
version = "0.1.0"
description = "AI-powered boat race prediction system using Claude API"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
boatrace = "boatrace_ai.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/boatrace_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
